Store client_id and oauth_state in AuthFlowContext
This tweaks the login code slightly to be more consistent with how it
handles the 3 OAuth-related parameters: `redirect_uri`, `client_id`, and
`state`.  These now get stored when initially creating the flow, and do
not need to be passed in again on requests to advance the flow.

The primary motivation for this change is to support "external" steps in
the login flow, where the client is redirected to an external server
which can perform authentication, before redirecting them back to their
final destination.  In order to do this, we need to store these 3 OAuth
parameters so that they can be looked up and used when finishing
authentication.  Previously the `client_id` and `state` fields were only
stored on the frontend while processing the `/auth/authorize` endpoint,
so their information gets lost if the user is redirected away from this
page.

As part of this change, we now also validate the redirect_uri when first
starting the login attempt, rather than only validating it after the
login is successful.  This also helps streamline the external
authentication flow process, since this validation has already been
performed before we redirect the user to the external authentication
server.